backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0          # async PostgreSQL driver
psycopg2-binary==2.9.9   # sync fallback / alembic

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Google Drive
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0

# CSV / file handling
python-multipart==0.0.9
aiofiles==23.2.1
